re-organize file structure
