+++ scavenger-client/src/components/App/App.js
