+++ scavenger-client/src/components/App/App.test.js
